requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
supabase>=2.0.0
google-generativeai>=0.3.0
bcrypt>=4.0.0
PyJWT>=2.8.0
pydantic[email]>=2.0.0
pygetwindow>=0.0.9
anthropic>=0.18.0  # optional, for real Claude API
